Create Number Guessing Game
A game in which the user must input a value, 1-100, and will receive a message whether they need to guess higher or lower, or if they guessed correctly. This program (in the C++ language) uses the while loop, if and else if statements. This program also incorporates the rand and srand function to ensure a new value every time the program is ran.